Validations

+++ app/Http/Controllers/ProjectController.php
@@ -34,7 +34,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -56,6 +56,10 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbiden'];
+        }
+
         return $this->repository->find($id);
     }
 
@@ -68,6 +72,10 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access forbiden'];
+        }
+
         return $this->service->update($request->all(), $id);
     }
 
@@ -79,6 +87,32 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access forbiden'];
+        }
+
         return $this->repository->delete($id);
     }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)) {
+            return true;
+        }
+
+        return false;
+    }
+
 }
